requirements.txt
MLB_StatsAPI>=1.9.0
Pillow>=10.0.1,<12.0.0
RGBMatrixEmulator>=0.13.5
tzlocal==4.2
bdfparser
./bullpen
./standings
./news
